main to design_documents

+++ src/main/java/usecase/coursePage/CoursePage.java
@@ -9,6 +9,9 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * A course page, that holds a course, related ratings and comment section(commentGraph)
+ */
 public class CoursePage implements Serializable {
 
     /**
@@ -24,43 +27,22 @@ public class CoursePage implements Serializable {
     private Course course;
     private List<String> instructors;
     private List<Rating> ratings;
-    private float averageScore;
+    private double averageScore;
     private CommentGraph commentGraph;
 
-
     /**
-     *    { Mathematics:  { Specialist: [Rating1, Rating2, Rating3]
-     *                       Major: [Rating1, Rating2, Rating3]
-     *                       Minor: [Rating1, Rating2, Rating3] }
-     *
-     *      Philosophy:   { Specialist: [Rating1, Rating2, Rating3]
-     *                       Major: [Rating1, Rating2, Rating3]
-     *                       Minor: [Rating1, Rating2, Rating3] }
-     *
-     *
-     *      Computer Science: { Specialist: [Rating1, Rating2, Rating3]
-     *                              Major: [Rating1, Rating2, Rating3]
-     *                              Minor: [Rating1, Rating2, Rating3] }
-     *
-     *    Constants for programs:
-     *
-     *
-     *
+     * Initializes a new coursePage
+     * @param course
+     * @param instructors
      */
-
-//    public float getRelativeRating(List<Rating> ratings) {
-//        HashMap<String, HashMap<String, List<Rating>>> programToRating = new HashMap<>();
-//        for (Rating r : ratings) {
-//            programToRating.put()
-//        }
-//    }
-
     public CoursePage(Course course, List<String> instructors) {
 
         this.course = course;
         this.instructors = instructors;
         this.ratings = new ArrayList<>();
     }
+
+    // Getters
 
     public Course getCourse() {
         return this.course;
@@ -74,7 +56,7 @@ public class CoursePage implements Serializable {
         return this.ratings;
     }
 
-    public float getAverageScore() {
+    public double getAverageScore() {
         return this.averageScore;
     }
 
@@ -102,7 +84,7 @@ public class CoursePage implements Serializable {
         this.ratings = ratings;
     }
 
-    public void setAverageScore(float AverageScore) {
+    public void setAverageScore(double AverageScore) {
         this.averageScore = AverageScore;
     }
